requirements.txt
beautifulsoup4
Flask
pandas
plotly
python-dotenv
Requests
gunicorn
mysql-connector-python
validate-email
Flask-Mail
weasyprint
pytest
